chore(geo): Update logging levels for geolocation process
This commit adjusts the logging levels for the geolocation process in `get_geo_position` function. The logging messages related to setting the desired accuracy, getting the geoposition asynchronously, and extracting details from the geoposition have been changed from `info!` to `trace!` to reduce noise in the logs for production use. An additional `info!` log has been added to output the current geoposition after it is successfully retrieved.